requirements.txt
streamlit
langchain
python-dotenv
langchain-openai
langchain-google-genai